Move com_init_wrapper to separate file
